Move onNetworkReady to chain.js and update the peers.js
