--- framework/src/modules/chain/submodules/index.js
@@ -1,3 +1,0 @@
-/**
- * @namespace modules
- */
